Refactor api server to a seperate file
