+++ 02-Calling-an-API/ApiServer.js
@@ -1,0 +1,44 @@
+const express = require("express");
+const morgan = require("morgan");
+const helmet = require("helmet");
+const jwt = require("express-jwt");
+const jwksRsa = require("jwks-rsa");
+const { join } = require("path");
+const cors = require("cors");
+const authConfig = require("./src/auth_config.json");
+
+const app = express();
+
+const port = process.env.SERVER_PORT || 3001;
+
+if (!authConfig.domain || !authConfig.audience) {
+  throw new Error(
+    "Please make sure that auth_config.json is in place and populated"
+  );
+}
+
+app.use(cors());
+app.use(morgan("dev"));
+app.use(helmet());
+app.use(express.static(join(__dirname, "build")));
+
+const checkJwt = jwt({
+  secret: jwksRsa.expressJwtSecret({
+    cache: true,
+    rateLimit: true,
+    jwksRequestsPerMinute: 5,
+    jwksUri: `https://${authConfig.domain}/.well-known/jwks.json`
+  }),
+
+  audience: authConfig.audience,
+  issuer: `https://${authConfig.domain}/`,
+  algorithm: ["RS256"]
+});
+
+app.get("/api/external", checkJwt, (req, res) => {
+  res.send({
+    msg: "Your access token was successfully validated!"
+  });
+});
+
+app.listen(port, () => console.log(`API server listening on port ${port}`));
