+++ 02-Calling-an-API/src/react-auth0-spa.js
@@ -1,7 +1,6 @@
 import React, { useState, useEffect, useContext } from "react";
 import createAuth0Client from "@auth0/auth0-spa-js";
 import config from "./auth_config.json";
-import history from "./utils/history";
 
 const DEFAULT_REDIRECT_CALLBACK = () =>
   window.history.replaceState({}, document.title, window.location.pathname);
